+++ app/monitoring/health_check.js
@@ -1,5 +1,6 @@
 var Step = require('step'),
-    _    = require('underscore');
+    _    = require('underscore'),
+    fs   = require('fs');
 
 function HealthCheck(metadataBackend, psqlClass) {
     this.metadataBackend = metadataBackend;
@@ -17,41 +18,22 @@ HealthCheck.prototype.check = function(username, query, callback) {
         };
 
     Step(
-        function getDBParams() {
-            startTime = Date.now();
-            self.metadataBackend.getAllUserDBParams(username, this);
+        function getManualDisable() {
+          fs.readFile(global.settings.disabled_file, this);
         },
-        function runQuery(err, dbParams) {
-            result.redis.ok = !err;
-            result.redis.elapsed = Date.now() - startTime;
-
-            if (err) {
-                throw err;
-            }
-
-            result.redis.count = Object.keys(dbParams).length;
-
-            var psql = new self.psqlClass({
-                host: dbParams.dbhost,
-                port: global.settings.db_port,
-                dbname: dbParams.dbname,
-                user: _.template(global.settings.db_user, {user_id: dbParams.dbuser}),
-                pass: _.template(global.settings.db_user_pass, {
-                    user_id: dbParams.dbuser,
-                    user_password: dbParams.dbpass
-                })
-            });
-
-            startTime = Date.now();
-            psql.query(query, this);
+        function handleDisabledFile(err, data) {
+          var next = this;
+          if (err) {
+            return next();
+          }
+          if (!!data) {
+            err = new Error(data);
+            err.http_status = 503;
+            throw err;
+          }
         },
-        function handleQuery(err, resultSet) {
-            result.postgresql.ok = !err;
-            if (!err) {
-                result.postgresql.elapsed = Date.now() - startTime;
-                result.postgresql.count = resultSet.rows.length;
-            }
-            callback(err, result);
+        function handleResult(err) {
+          callback(err, result);
         }
     );
 };